requirements.txt
git+https://github.com/TelegramPlayground/pyrogram
git+https://github.com/TelegramPlayGround/pyrogram-tgcrypto
norsodikin
python-dotenv
